Migrate Hospital type to own file.
