+++ frontend/Components/G0vbar.jsx
@@ -1,7 +1,7 @@
 // @flow
 import * as React from 'react';
 // $FlowIgnore[cannot-resolve-module] This is how ParcelJS imports assets
-import favicon from 'url:../assets/g0v.png'; // eslint-disable-line
+import favicon from 'url:../../assets/g0v.png'; // eslint-disable-line
 
 import type { Locale } from '../Types/Locale';
