+++ frontend/Components/Table.jsx
@@ -6,17 +6,8 @@ import strings from '../Strings/Table.json';
 import type { Locale } from '../Types/Locale';
 import type { Availability } from '../Types/Availability';
 import type { Location } from '../Types/Location';
+import type { Hospital } from '../Types/Hospital';
 
-export type Hospital = {|
-  hospitalId: Number,
-  name: string,
-  location: Location,
-  department: string,
-  phone: string,
-  address: string,
-  website: string,
-  availability: Availability,
-|};
 
 export default function Table(props: { rows: Array<Hospital>, locale: Locale }): React.Node {
   const { rows, locale } = props;
